# (Context).get: speedup