`fmt.Sprintf` is slow. Let's use string concatenation instead, and avoid
looking up level label twice.

Benchmark shows 3x improvement in performance, and less allocations.
```
name          old time/op    new time/op    delta
ContextGet-4     277ns ± 1%      93ns ± 1%  -66.55%  (p=0.016 n=4+5)

name          old alloc/op   new alloc/op   delta
ContextGet-4      128B ± 0%       64B ± 0%     ~     (p=0.079 n=4+5)

name          old allocs/op  new allocs/op  delta
ContextGet-4      5.00 ± 0%      1.00 ± 0%     ~     (p=0.079 n=4+5)
```
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>