@kolyshkin 2021-09-09
> I am not sure this is called that often.

It's not, but the fix is trivial, too.